server/api-v1/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
Django==4.1
django-cognito-jwt==0.0.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-jwt==1.19.2
ecdsa==0.18.0
idna==3.3
install==1.3.5
jwt==1.3.1
Naked==0.1.31
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pycrypto==2.6.1
PyJWT==2.4.0
python-jose==3.3.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
rsa==4.9
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.11
